--- PennMobile/GSR-Booking/Model/LocationsHandler.swift
@@ -1,24 +1,0 @@
-//
-//  GSRLocationsHandler.swift
-//  GSR
-//
-//  Created by Yagil Burowski on 15/09/2016.
-//  Copyright © 2016 Yagil Burowski. All rights reserved.
-//
-
-import Foundation
-
-class LocationsHandler {
-    static func getLocations() -> [GSRLocation] {
-        return [GSRLocation(name: "Weigle", code: 1722, path: "/booking/wic"),
-                GSRLocation(name: "VP GSR", code: 1799, path: "/booking/vpdlc"),
-                GSRLocation(name: "Lippincott", code: 1768, path: "/booking/lippincott"),
-                GSRLocation(name: "Edu Commons", code: 848, path: "/booking/educom"),
-                GSRLocation(name: "Levin Building", code: 13489, path: "/booking/levin"),
-                GSRLocation(name: "VP Sem. Rooms", code: 4409, path: "/booking/seminar"),
-                GSRLocation(name: "Lippincott Sem. Rooms", code: 2587, path: "/booking/lippseminar"),
-                GSRLocation(name: "Glossberg Recording Room", code: 1819, path: "/booking/glossberg"),
-                GSRLocation(name: "Biomedical Lib.", code: 505, path: "/booking/biomed")
-        ]
-    }
-}
