+++ PennMobile/GSR-Booking/Model/StudySpace.swift
